requirements.txt
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.38.2
numpy
python-dotenv
redis==3.5.3
networkx==2.6.3
mysqlclient